added correct slicing

+++ crates/polars-plan/src/dsl/function_expr/array.rs
@@ -7,7 +7,7 @@ use polars_core::prelude::SortOptions;
 #[cfg_attr(feature = "dsl-schema", derive(schemars::JsonSchema))]
 pub enum ArrayFunction {
     Length,
-    Slice(i64, usize),
+    Slice(i64, i64),
     Min,
     Max,
     Sum,
